+++ apps/dine-pleiepenger/src/utils/søknadUtils.ts
@@ -1,7 +1,13 @@
 import dayjs from 'dayjs';
 import timezone from 'dayjs/plugin/timezone';
 import utc from 'dayjs/plugin/utc';
-import { Søknad } from '../types/Søknad';
+import {
+    PleiepengerEndringsmelding,
+    PleiepengerEttersendelse,
+    Pleiepengesøknad,
+    Søknad,
+    Søknadstype,
+} from '../types/Søknad';
 
 require('dayjs/locale/nb');
 
@@ -16,4 +22,18 @@ export const formatSøknadOpprettetDato = (date: Date) => {
 export const sortSøknadEtterOpprettetDato = (a: Søknad, b: Søknad, desc: boolean = true): number => {
     const direction = desc ? 1 : -1;
     return new Date(a.opprettet) < new Date(b.opprettet) ? direction : direction * -1;
+};
+
+export const getPleiepengesøknader = (søknader: Søknad[]): Pleiepengesøknad[] => {
+    return søknader.filter((søknad) => søknad.søknadstype === Søknadstype.PP_SYKT_BARN) as Pleiepengesøknad[];
+};
+export const getEndringsmeldinger = (søknader: Søknad[]): PleiepengerEndringsmelding[] => {
+    return søknader.filter(
+        (søknad) => søknad.søknadstype === Søknadstype.PP_SYKT_BARN_ENDRINGSMELDING,
+    ) as PleiepengerEndringsmelding[];
+};
+export const getEttersendelser = (søknader: Søknad[]): PleiepengerEttersendelse[] => {
+    return søknader.filter(
+        (søknad) => søknad.søknadstype === Søknadstype.PP_ETTERSENDELSE,
+    ) as PleiepengerEttersendelse[];
 };