+++ apps/endringsmelding-pleiepenger/.storybook/main.ts
@@ -1,6 +1,6 @@
 const config = {
     stories: ['../src/storybook/**/*.stories.tsx'],
-    addons: ['@storybook/addon-essentials', '@storybook/addon-a11y', 'storybook-formik/register'],
+    addons: ['@storybook/addon-essentials', '@storybook/addon-a11y'],
     framework: {
         name: '@storybook/react-vite',
         options: {},