+++ apps/endringsmelding-pleiepenger/src/app/api/endpoints/sakerEndpoint.ts
@@ -1,3 +1,4 @@
+import { isUnauthorized } from '@navikt/sif-common-core-ds/src/utils/apiUtils';
 import { getEnvironmentVariable } from '@navikt/sif-common-core-ds/src/utils/envUtils';
 import { isK9FormatError, K9Format, K9FormatArbeidstid, K9Sak, UgyldigK9SakFormat } from '@types';
 import {
@@ -8,9 +9,10 @@ import {
     maskString,
     parseK9Format,
 } from '@utils';
+import { isAxiosError } from 'axios';
+import { verifyK9Format } from '../../utils/verifyk9Format';
 import api from '../api';
 import { ApiEndpointInnsyn } from './';
-import { verifyK9Format } from '../../utils/verifyk9Format';
 
 export type K9SakResult = K9Sak | UgyldigK9SakFormat;
 
@@ -37,7 +39,13 @@ const sakerEndpoint = {
     fetch: async (): Promise<{ k9Saker: K9SakResult[]; eldreSaker: K9SakResult[] }> => {
         const endringsperiode = getTillattEndringsperiode(getEndringsdato());
         try {
+            appSentryLogger.logInfo(`sakerEndpoint.fetch saker`);
             const { data } = await api.innsyn.get<K9Format[]>(ApiEndpointInnsyn.sak);
+            try {
+                appSentryLogger.logInfo(`sakerEndpoint.fetch length: ${data.length}`);
+            } catch (error) {
+                appSentryLogger.logInfo(`sakerEndpoint.fetch noLength, ${typeof data}`);
+            }
             const k9Saker: K9SakResult[] = [];
             const eldreSaker: K9SakResult[] = [];
             data.forEach((sak) => {
@@ -72,6 +80,12 @@ const sakerEndpoint = {
             });
             return Promise.resolve({ k9Saker, eldreSaker });
         } catch (error) {
+            if (isAxiosError(error) && !isUnauthorized(error)) {
+                appSentryLogger.logInfo(`sakerEndpoint.fetch failed - ${error.message}`);
+            } else {
+                appSentryLogger.logInfo('sakerEndpoint.fetch failed - unauthorized');
+            }
+            appSentryLogger.logInfo('sakerEndpoint.fetch failed - something');
             return Promise.reject(error);
         }
     },