+++ apps/endringsmelding-pleiepenger/src/app/api/fetchInitialData.ts
@@ -37,7 +37,12 @@ export const fetchInitialData = async (
 }> => {
     const [søker, k9sakerResult] = await Promise.all([søkerEndpoint.fetch(), sakerEndpoint.fetch()]);
 
+    if (k9sakerResult.k9Saker.length === 0 && k9sakerResult.eldreSaker.length === 0) {
+        appSentryLogger.logInfo('fetchInitialData.ingenSaker');
+    }
+
     const handleInitialDataError = (error: any) => {
+        appSentryLogger.logInfo(`handleInitialDataError: ${error?.message}`);
         if (isSøknadInitialDataErrorState(error)) {
             return Promise.reject({
                 ...error,
@@ -151,6 +156,7 @@ const kontrollerSaker = (
 const kontrollerTilgang = async (k9saker: K9Sak[], tillattEndringsperiode: DateRange): Promise<boolean> => {
     const resultat = tilgangskontroll(k9saker, tillattEndringsperiode);
     if (resultat.kanBrukeSøknad) {
+        appSentryLogger.logInfo(`kontrollerTilgang.kanBrukeSøknad`);
         return Promise.resolve(true);
     }
     if (getEnvironmentVariable('DEBUG') === 'true') {
@@ -162,8 +168,7 @@ const kontrollerTilgang = async (k9saker: K9Sak[], tillattEndringsperiode: DateR
                     sak: maskK9Sak(k9saker[0]),
                 }),
             );
-        }
-        if (k9saker.length > 1) {
+        } else {
             appSentryLogger.logInfo(
                 'IkkeTilgangSakInfo',
                 JSON.stringify({
@@ -172,6 +177,7 @@ const kontrollerTilgang = async (k9saker: K9Sak[], tillattEndringsperiode: DateR
             );
         }
     }
+    appSentryLogger.logInfo(`kontrollerTilgang.kanIkkeBrukeSøknad: ${resultat.årsak}`);
     return Promise.reject(getKanIkkeBrukeSøknadRejection(resultat.årsak, resultat.ingenTilgangMeta));
 };