+++ apps/omsorgspengerutbetaling-soknad/src/app/søknad/steps/fravær/FraværStepInfo.tsx
@@ -1,4 +1,4 @@
-import { FormattedMessage, useIntl } from 'react-intl';
+import { FormattedMessage, IntlShape } from 'react-intl';
 import ExpandableInfo from '@navikt/sif-common-core-ds/src/components/expandable-info/ExpandableInfo';
 import intlHelper from '@navikt/sif-common-core-ds/src/utils/intlUtils';
 import SifGuidePanel from '@navikt/sif-common-core-ds/src/components/sif-guide-panel/SifGuidePanel';
@@ -18,8 +18,7 @@ const IntroVeileder = () => {
     );
 };
 
-const Tidsbegrensning = (delvisdag?: boolean) => {
-    const intl = useIntl();
+const Tidsbegrensning = (intl: IntlShape, delvisdag?: boolean) => {
     return (
         <ExpandableInfo title={intlHelper(intl, 'step.fravaer.info.ikkeHelg.tittel')}>
             {delvisdag && <FormattedMessage id="step.fravaer.delvisdag.info.ikkeHelg.tekst" />}